* commit '4c2d377e0d1d4e62a05482e0bd2c23173042f26c':
  NEXT-2593 - Move StateMachineStates to Checkout domains
  NEXT-2593 - Use available shippingMethod in StorefrontTests
  NEXT-2593 - Remove Default locale
  NEXT-2593 - Remove Default localeIds
  NEXT-2593 - Remove locale defautls were they are not necessary
  NEXT-2593 - Remove snippetId defaults
  NEXT-2593 - Remove shipping method default
  NEXT-2593 - Move StateMachine constant to own classes
  NEXT-2593 - Remove defaults only used in Migrations

+++ src/Storefront/Test/Page/Checkout/ConfirmPageTest.php
@@ -5,7 +5,6 @@ namespace Shopware\Storefront\Test\Page\Checkout;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Checkout\Payment\PaymentMethodEntity;
 use Shopware\Core\Checkout\Shipping\ShippingMethodEntity;
-use Shopware\Core\Defaults;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
@@ -57,25 +56,28 @@ class ConfirmPageTest extends TestCase
         /** @var EntityRepositoryInterface $shippingMethodRepository */
         $shippingMethodRepository = $this->getContainer()->get('shipping_method.repository');
         $shippingMethodRuleRepository = $this->getContainer()->get('shipping_method_rule.repository');
-        /** @var ShippingMethodEntity $shippingMethod */
-        $shippingMethod = $shippingMethodRepository->search(new Criteria([Defaults::SHIPPING_METHOD]), $context->getContext())->get(Defaults::SHIPPING_METHOD);
+        $shippingMethods = $shippingMethodRepository->search(new Criteria(), $context->getContext())->getEntities();
 
         $ruleToDelete = [];
 
-        foreach ($shippingMethod->getAvailabilityRuleIds() as $availabilityRuleId) {
-            $ruleToDelete[] = [
-                'shippingMethodId' => Defaults::SHIPPING_METHOD,
-                'ruleId' => $availabilityRuleId,
-            ];
+        /** @var ShippingMethodEntity $shippingMethod */
+        foreach ($shippingMethods as $shippingMethod) {
+            foreach ($shippingMethod->getAvailabilityRuleIds() as $availabilityRuleId) {
+                $ruleToDelete[] = [
+                    'shippingMethodId' => $shippingMethod->getId(),
+                    'ruleId' => $availabilityRuleId,
+                ];
+            }
         }
 
-        $shippingMethodRuleRepository->delete($ruleToDelete, $context->getContext());
+        if (\count($ruleToDelete) > 0) {
+            $shippingMethodRuleRepository->delete($ruleToDelete, $context->getContext());
+        }
 
         /** @var CheckoutConfirmPageLoadedEvent $event */
         $event = null;
         $this->catchEvent(CheckoutConfirmPageLoadedEvent::NAME, $event);
 
-        $context = $this->createCheckoutContextWithNavigation();
         /** @var CheckoutConfirmPage $page */
         $page = $this->getPageLoader()->load($request, $context);
